% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicago_community_areas_map.R
\docType{data}
\name{chicago_community_areas_map}
\alias{chicago_community_areas_map}
\title{Chicago Community Areas (Current) GeoJSON}
\format{
GeoJSON spatial vector layer.
}
\usage{
chicago_community_areas_map
}
\description{
A GeoJSON file containing boundary polygons for community areas of Chicago, Illinois (current dataset).
}
\examples{
\dontrun{
library(sf)
library(ggplot2)
ggplot(chicago_community_areas_map) +
    geom_sf() +
    coord_sf() +
    theme_minimal()
}
}
\keyword{datasets}
