% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_county_inc_map.R
\docType{data}
\name{cook_county_inc_map}
\alias{cook_county_inc_map}
\title{Cook County Incorporated Areas GeoJSON}
\format{
GeoJSON spatial vector layer.
}
\usage{
cook_county_inc_map
}
\description{
A GeoJSON spatial file containing boundaries for incorporated areas in Cook County, Illinois.
}
\examples{
\dontrun{
library(sf)
library(ggplot2)
ggplot(cook_county_inc_map) +
    geom_sf() +
    coord_sf() +
    theme_minimal()
}
}
\keyword{datasets}
